Extract active-replay identity handling into a mixin (WS5 seam)
Move the 19 methods that match stored rows against ignore patterns,
cache and copy active-replay messages, restore externalized
ingest-payload placeholders, and prove/redact persisted-output
identities during replay -- verbatim out of engine.py into a new
ReplayIdentityMixin in replay_identity.py. LCMEngine mixes it in
before ContextEngine, so `self` still resolves every reference; call
sites, shared state, and method names are unchanged.

engine.py -345 lines. The 12 imports the block solely used move to
replay_identity.py and are dropped from engine.py.

Behaviour-neutral: the moved method bodies are byte-identical to
main (AST-slice compare, 0 diffs). ruff clean; full suite
1575 passed / 12 xfailed (unchanged from the v0.19.0 baseline).
